Criar as Entities e Repositories das classes Movimento, Operacao e Pagamento

+++ src/main/java/app/models/entities/Movimento.java
@@ -3,7 +3,6 @@ package app.models.entities;
 import lombok.AllArgsConstructor;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.sql.Date;
 
 @AllArgsConstructor
@@ -11,7 +10,7 @@ public class Movimento {
     private final Integer id;
     private final String cpf;
     private final Integer idFonte;
-    private final BigInteger numeroContrato;
+    private final Integer numeroContrato;
     private final Date dataVencimentoParcela;
     private final Integer quantidadeParcelasVencer;
     private final Integer quantidadeParcelasPagar;
