+++ src/main/java/app/models/entities/Pagamento.java
@@ -3,7 +3,6 @@ package app.models.entities;
 import lombok.AllArgsConstructor;
 
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.sql.Date;
 
 @AllArgsConstructor
@@ -11,7 +10,7 @@ public class Pagamento {
     private final Integer id;
     private final String cpf;
     private final Integer idFonte;
-    private final BigInteger numeroContrato;
+    private final Integer numeroContrato;
     private final Date dataVencimentoParcela;
     private final Date dataPagamentoParcela;
     private final BigDecimal valorPago;
